cleanup - move all application configuration  from api into configuration submodule
