--- api/src/main/kotlin/github/com/hukuta94/delivery/startup/configuration/infrastructure/adapter/inmemory/InMemoryCourierRepositoryConfiguration.kt
@@ -1,12 +1,0 @@
-package github.com.hukuta94.delivery.startup.configuration.infrastructure.adapter.inmemory
-
-import github.com.hukuta94.delivery.infrastructure.adapter.inmemory.CourierInMemoryRepository
-import org.springframework.context.annotation.Bean
-import org.springframework.context.annotation.Configuration
-
-@Configuration
-open class InMemoryCourierRepositoryConfiguration {
-
-    @Bean
-    open fun courierRepository() = CourierInMemoryRepository()
-}
