--- api/src/main/kotlin/github/com/hukuta94/delivery/startup/configuration/infrastructure/adapter/orm/OrmOrderRepositoryConfiguration.kt
@@ -1,20 +1,0 @@
-package github.com.hukuta94.delivery.startup.configuration.infrastructure.adapter.orm
-
-import github.com.hukuta94.delivery.infrastructure.adapter.orm.repository.domain.OrderJpaRepository
-import github.com.hukuta94.delivery.infrastructure.adapter.orm.repository.domain.OrmOrderRepositoryAdapter
-import github.com.hukuta94.delivery.infrastructure.adapter.orm.repository.event.OrmOutboxEventRepositoryAdapter
-import org.springframework.context.annotation.Bean
-import org.springframework.context.annotation.Configuration
-
-@Configuration
-open class OrmOrderRepositoryConfiguration {
-
-    @Bean
-    open fun orderRepository(
-        orderJpaRepository: OrderJpaRepository,
-        outboxRepository: OrmOutboxEventRepositoryAdapter,
-    ) = OrmOrderRepositoryAdapter(
-        orderJpaRepository = orderJpaRepository,
-        outboxRepository = outboxRepository,
-    )
-}
