+++ configuration/src/main/kotlin/github/com/hukuta94/delivery/DeliveryApplication.kt
@@ -1,5 +1,6 @@
 package github.com.hukuta94.delivery
 
+import github.com.hukuta94.delivery.configuration.DeliveryApplicationConfiguration
 import org.springframework.boot.runApplication
 
 fun main(args: Array<String>) {
