+++ configuration/src/main/kotlin/github/com/hukuta94/delivery/configuration/api/adapter/http/RestControllerConfiguration.kt
@@ -1,0 +1,32 @@
+package github.com.hukuta94.delivery.configuration.api.adapter.http
+
+import org.springframework.context.annotation.Bean
+import org.springframework.context.annotation.Configuration
+import org.springframework.context.annotation.Import
+import org.springframework.web.cors.CorsConfiguration
+import org.springframework.web.cors.UrlBasedCorsConfigurationSource
+import org.springframework.web.filter.CorsFilter
+import org.springframework.web.servlet.config.annotation.EnableWebMvc
+
+@EnableWebMvc
+@Configuration
+@Import(
+    OrderRestConfiguration::class,
+    CourierRestConfiguration::class,
+)
+open class RestControllerConfiguration {
+    @Bean
+    open fun corsFilter(): CorsFilter {
+        val config = CorsConfiguration().apply {
+            allowCredentials = true
+            addAllowedOrigin("http://localhost:8086")
+            addAllowedHeader("*")
+            addAllowedMethod("GET")
+        }
+        val source = UrlBasedCorsConfigurationSource().apply {
+            registerCorsConfiguration("/**", config)
+        }
+
+        return CorsFilter(source)
+    }
+}
