+++ configuration/src/main/kotlin/github/com/hukuta94/delivery/configuration/core/application/event/ApplicationEventsConfiguration.kt
@@ -1,0 +1,11 @@
+package github.com.hukuta94.delivery.configuration.core.application.event
+
+import org.springframework.context.annotation.Configuration
+import org.springframework.context.annotation.Import
+
+@Configuration
+@Import(
+    DomainEventsConfiguration::class,
+    IntegrationEventsConfiguration::class,
+)
+open class ApplicationEventsConfiguration
