+++ configuration/src/main/kotlin/github/com/hukuta94/delivery/configuration/infrastructure/InfrastructureConfiguration.kt
@@ -1,0 +1,13 @@
+package github.com.hukuta94.delivery.configuration.infrastructure
+
+import github.com.hukuta94.delivery.configuration.infrastructure.adapter.kafka.KafkaProducerConfiguration
+import github.com.hukuta94.delivery.configuration.infrastructure.adapter.orm.OrmRepositoryConfiguration
+import org.springframework.context.annotation.Configuration
+import org.springframework.context.annotation.Import
+
+@Configuration
+@Import(
+    OrmRepositoryConfiguration::class,
+    KafkaProducerConfiguration::class,
+)
+open class InfrastructureConfiguration
