+++ configuration/src/main/kotlin/github/com/hukuta94/delivery/configuration/infrastructure/adapter/orm/OrmCourierRepositoryConfiguration.kt
@@ -1,0 +1,20 @@
+package github.com.hukuta94.delivery.configuration.infrastructure.adapter.orm
+
+import github.com.hukuta94.delivery.infrastructure.adapter.orm.repository.domain.CourierJpaRepository
+import github.com.hukuta94.delivery.infrastructure.adapter.orm.repository.domain.OrmCourierRepositoryAdapter
+import github.com.hukuta94.delivery.infrastructure.adapter.orm.repository.event.OrmOutboxEventRepositoryAdapter
+import org.springframework.context.annotation.Bean
+import org.springframework.context.annotation.Configuration
+
+@Configuration
+open class OrmCourierRepositoryConfiguration {
+
+    @Bean
+    open fun courierRepository(
+        courierJpaRepository: CourierJpaRepository,
+        outboxRepository: OrmOutboxEventRepositoryAdapter,
+    ) = OrmCourierRepositoryAdapter(
+        courierJpaRepository = courierJpaRepository,
+        outboxRepository = outboxRepository,
+    )
+}
